In order to get a more snappy feeling from the CLI help system, only load the information when it is actually referenced, rather than loading everything up-front.

The performance impact is not huge compared to `main`, but this opens up for adding more data without having to pay the cost in time up-front, for things like rules and api types info (see #14227)

On `main` 1e1fdabfc:
```
$ hyperfine --warmup=2 --runs=10 './pants help-advanced docker'
Benchmark 1: ./pants help-advanced docker
  Time (mean ± σ):      1.648 s ±  0.023 s    [User: 0.758 s, System: 0.396 s]
  Range (min … max):    1.624 s …  1.695 s    10 runs
 
```

This branch:
```
$ hyperfine --warmup=2 --runs=10 './pants help-advanced docker'
Benchmark 1: ./pants help-advanced docker
  Time (mean ± σ):      1.497 s ±  0.018 s    [User: 0.758 s, System: 0.386 s]
  Range (min … max):    1.481 s …  1.541 s    10 runs
```

And for `help-all` the difference is negligible:
```
# main
$ hyperfine --warmup=2 --runs=10 './pants help-all'
Benchmark 1: ./pants help-all
  Time (mean ± σ):      1.722 s ±  0.027 s    [User: 0.766 s, System: 0.387 s]
  Range (min … max):    1.691 s …  1.779 s    10 runs

# This branch
$ hyperfine --warmup=2 --runs=10 './pants help-all'
Benchmark 1: ./pants help-all
  Time (mean ± σ):      1.711 s ±  0.057 s    [User: 0.764 s, System: 0.388 s]
  Range (min … max):    1.631 s …  1.799 s    10 runs
```

There are a few places with "ugly" lazy loading code for the rule infos, which will be refactored and cleaned up in #14227